Rewrites documentation for rotate3_inertia and renames to relevant functions.
The rotate3_inertia's documentation defined the rotation matrix as v_a = R *
v_b but what it actually returned seemed to be from the definition v_b = R *
v_a. I updated the documentation to reflect this and altered the
Human.inertia_transformed to return what it's documentation says it returns.
Finally I renamed rotate3_inertia to rotate_inertia and switched the order of
the arguments and I renamed Human.inertia_transformed to
Human.transform_inertia so that the name is a verb indicating what the method
does rather than what it returns. Both of these have deprecation warnings so
the old names can be used until Yeadon 2.0. I also moved the deprecation
warning to it's own module so that it can be more easily used in all the
modules with no chance of circular dependencies.