Fixes #84 : Purges vb = R * va rotation definitions.
We had two definitions of rotation matrices in the package and things were
confusing. I removed all traces of the definition that was not being used in
the yeadon software itself. This leaves us with all rotation matrices, R,
defined as va = R * vb where reference frame B is rotated relative to reference
frame A and va is v expressed in A and vb is v expressed in B.