replaced yaml.load by yaml.safe_load
yaml.load was deprecated in PyYAML 5.1+ and has now changed the API.
This causes an exception to be raised asking to provide a Loader in the
current release of yeadon.

yaml.safe_load is present in all release of PyYAML and is sufficient for
yeadon's purpose.

For more information, see PyYAML's wiki entry.
[https://github.com/yaml/pyyaml/wiki/PyYAML-yaml.load(input)-Deprecation]